requirements.txt
streamlit==1.40.0
selenium==4.26.1
python-dotenv==1.0.1
fake-useragent==1.5.1